Moving to TypeScript
I've been interested for the past year or so in how this library works internally. So, I felt like this would be the best way to find out about that, by moving it to TypeScript!

I think I can do a lot of other beneficial things for the library too, like bringing native ESM support, and using cross-platform standard APIs that are supported by both the browser and NodeJS nowadays. This could possibly make things a bit lighter in terms of the build step, as it won't need to rely on having polyfills to work the same in the browser. It may not be doing this too much, as I don't know too much about how everything works completely yet, but this has seemed to help with other similar projects I have done this with. It's a very relieving process to do things like this! :)