Near Type-Safe
This is the first iteration at moving things over, and it went very well! This was from the last two hours or so.

There were a few parts I had to leave for future changes, like having to do with Object index signature accessing, and things like that.

This also moves everything to being ESM-compatible too! Super awesome. I still have to add the file extensions to the import statements too though, lemme do that next real quick.