Modern Unit Test Tooling
Trying to update the unit tests to work with the new TypeScript and ESM setup.

I'm not sure what I want to do for this yet, I'm trying out a few different things to start with. There's a few different options:

- Stay with Jest, keep using Babel (maybe in conjunction with ts-jest or something like that, whatever will work)
- Vitest (I think that uses esbuild, if anything)
- NodeJS built in new test runner (I really like this option, it doesn't require a build step just for testing)

I tried some of this yesterday, but I started fresh again tonight when looking at the original changes I made to the codebase.