Unit Testing Migration (Very Early, experimental)
There's a lot of code to go through for theses tests, and it's still not working very well for me on trying to understand how to do testing with TypeScript and a light test runner. This is new to me, it's a bit above my head still. That's what I like about it though, means I think I'm on the right track to learning something new.

I was also going to remove some old configuration files at the root of the project, but I don' think I'm ready to do that until I have the new build and testing processes all set up and ready to go. Once that's the case, then I think I can remove those, as well as the old project dependencies.

I guess I'm trying to make this project a bit more lightweight like I tend to do with my own projects.

And I had to change the tsconfig to allow the same type checking inside of the test folder. This means I'll have to pass custom rootDir and outDir flags to the tsc build step itself, rather than listing them from inside of the tsconfig. That's actually a very great workaround, and it may also actually be able to help me in some of my other projects!