TS-Jest
Got Jest working with the new TypeScript codebase! I'm very happy that a lot of the tests are still passing, so at least the ESM part is working, that was the biggest hurdle in getting this to work.

This setup is very hacky, with all of the dependencies that it currently requires. I really don't like how much spaghetti code it needs to work with simple ESM TypeScript, how funky haha (Of course, these tools have been around a while before this 'newer' spec, but you'd think that things like this would be more of a vanilla experience to set up by now).

So here's the list of things going on. For simply building the source of the library, I'm only using `tsc`. To *test* the library (this is just for testing, get ready), I'm using Jest. To read the source, that uses Babel. Even with that, it isn't able to run TypeScript. So, I have to use TS-Jest, which depends on both `tsc` and Jest. Neither of these tools (excluding `tsc`) can handle having `.js` extension resolution in TS files, so I had to install a resolver plugin for Jest, which I think resolves the `.js` extension to the actual `.ts` one, since Jest doesn't know, or know how to do that. And because Jest (TS-Jest) doesn't support bare ESM, I think somewhere in there it's compiling the code to CommonJS still, since that's only what it can run? At least I think it's something like that. I don't know. But hey, it's working now! Let's cheers to improving this from here :)

I originally moved the tests back to the `__tests__` and `__mocks__` folders because I thought they had to be there in order for Jest to find them, but turns out the original, new folder structure works perfectly afterall! I'm really glad I looked into seeing if that would still work. I didn't even have to add a path configuration to Jest's config to make it work. Epic.

Next I'm going to work on the unused project dev dependencies, now that things pretty much only use TypeScript and Jest. I'm not sure how I want to handle the React Native and Node file opener tests/typings yet, that's been a hold back since the migration, as I'm not sure how those are fully supposed to work or be used. Rather than removing it, I wanted to leave it as long as possible to figure out how to keep it around, or if I should move it to a new implementation down the road, once I start to modernize the actual codebase.

I'm going to work on adding Promises to the codebase, that's gonna be a huge plus/level up :)