Jest Pure ESM
Awesome!!! Got Jest working with pure ESM and TypeScript. The only hold up is getting mocking working with the fs and xhr2 modules.

This involved adjusting TS-Jest to compile to pure ESM (which looks like it wasn't doing with my previous config for it), and it also requires running Jest using Node's experimental ESM VM modules flag.

Big thanks to the response from RimuTec on StackOverflow, and the link to their repo, which shows a working copy with this setup.

Once I got ESM running, this prevented the Jest namespace from being accessible from the module's global scope, so you have to manually import it again in order to access it (yay, I didn't like relying on that original global-based setup :) It's more explicit with this now, too).

https://stackoverflow.com/questions/72432973/ts-jest-cannot-use-top-level-await-even-with-correct-tsconfig
https://github.com/RimuTec/express-testing
https://jestjs.io/docs/ecmascript-modules