Deep Types
Stengthened the internal use of types within the library! I remembered that I am using the DefinitelyTyped definitions for jsmediatags in other projects of mine, so I decided to look into what types those types define differently over what the library is currently doing.

https://www.npmjs.com/package/@types/jsmediatags

I merged in some of the nice things that those build on top of the regular definitions, and it helped straighten out a few things in the codebase that didn't have any strict types for previously. This helped out a bunch actually! It's going to help with moving to Promises a lot easier, as now the error handling is typed out, rather than just being an `Object` with various unknown keys, passed through an `any` type. Now the tag shortcuts have types too, which was previously not described with the library's original Flow types.