Promise APIs
Moved all of the library over from the original callbacks syntax to Promises!

Unfortunately not all of the tests are passing yet, so I think somewhere in the code, there must be a case of leaving a Promise floating somewhere, as the tests appear to time out after more than a minute or so. The first few files are passing no problem though, so I think it may just be an isolated case somewhere that I haven't handled correctly yet. Everything is also type-safe still though, so that's great!

These are a lot of changes, and it's all mostly working, so I wanted to commit everything I have so far, since it's a great step forward.

A lot of the library uses async code for some reason, in parts that I wouldn't expect it to need. I'm not sure what that's all about, but I think figuring out what that's about, and hopefully slimming it down could really make it faster. I think doing that alongside using native TypedArrays instead of plain number arrays will also be a possible performance gain.

https://stackoverflow.com/questions/33562284/how-do-i-catch-thrown-errors-with-async-await

I'm not totally sure if I'm doing the async code correctly with the async-await code, so finding this post was a result of trying to learn more about if I am still testing things right.

https://stackoverflow.com/questions/72531324/jests-runallticks-doesnt-completely-exhaust-the-micro-task-queue