Lib-Return-Types
Simplified the return type declarations for various functions and methods, as now they use the inheritance of the parent class it's based on to provide the type checking for that method. I like this because it's a fair amount more dynamic, and it still provides parent-to-child method type validation, because of the `override` keyword handling. Now I'm going to try and do this with the public-private/protected state for properties and methods, and also the abstract syntax, if that will actually help in this case. I'm not sure if the current API inheritance stack is following something which can be labeled using the `abstract` syntax, I think some of the inheritance levels aren't quite solidly established between classes.