File Retrieving Readers (No more?)
Deciding to go to the next ideas of things, and remove the Node, XHR, and React Native file readers. I want to do this in favor of moving the library over to being more functionally structured, to where you pass in the direct data you want to open yourself, and the library handles it there. I don't want it to be responsible for the data retrieving stages, too. That's the same ideas I had with NBTify, you as the user are up to get the data from the source, then you can pipe it into the library, which is solely responsible for working with the data itself. I like this approach!