Formatting + Even Stricter Testing
Can't quite get the testing set up still. It's definitely really close now though! :D I think it just has to do with the CJS / ESM crossover that you have to deal with in Jest. That's one of my reasons for wanting to move away from Jest. Not sure how to do that quite yet though, so trying ti figure at least something out in the meantime seems like the safer route. More to learn about too :)

'The Scambot Holiday Special'

I was going to go back to before I removed the 'File Retrieving Readers', but I realized that it wouldn't help out too much. I was going to do that to help the main library with hopefully migrating to this version, but I think it will be holding it back from doing what this rework is trying to do in the first place. I want to help readjust where the inputs and outputs go for this library, so it's more dependency-injection-based. Then the library itself only has to worry about the internal logic it needs to actually read from files, rather than also implementing the IO parts, too.

Did some formatting I've been meaning to do for a while also. And for the `@ts-ignore` statements, I removed them now that the library is more strongly-typed itself. I want to use it now to debug the tests to figure out how to set them up correctly with the new changes. Glad I removed the ignore statements, because I found a few things I missed with the reader removals from the previous^ commit.