TSConfig Build v. Dev
This is essentially what I've been doing manually myself the last while. This is key to getting the correct IDE type checking for both the source and the tests, together. So I thought it would be a must have to enable this without knowing the need to edit the TSConfig just to get editor checking working, compared to just the build step. Now they are separate concerns, and they work together.