Class Visibility Modifiers
Added some additional scoping to the class properties and methods. Now it's more explicit as to what is used where! This was actually really efficient to run alongside `tsc -w`, as I could see if something couldn't be marked as private or protected, because it was being used in a test or something like that.