Concat to Spread
Moving the classic conat array setup calls over to using the spread operator instead! This will help with moving to TypedArrays also :D

Had the no-check call in the main test file for the last few commits, wasn't going to commit it because I was hoping to remove all of the uses of `// @ts-expect-error`, but I think doing all of them will be for another day. Most of the remaining type issues are because of object key indexing for metadata properties and such.