Proper TS Migration - CommonJS
Going to go back to the ground level again to start this TS conversion.

I recently learned with some other TS conversions that I was doing before this one (this was pretty much my first), I think I was doing some of the things in the wrong order.

It's absolutely crucial to preserve the state of your tests while they are still working, when moving code over. This is only going to be of help for when you change things later.

I should have waited longer before moving to pure ESM, and should have done the TS part first, sticking with CommonJS. Once things are being tested accordingly in CJS, I think it's not too far to ensure that moving to ESM isn't something too far off. And coming back to this project, I just realized how much is already done actually! Everything was written in Flow, which is very close to TS, so not much work has to be done by me on that front.

Listening to the bonus tracks from Empath, I'm really getting into it this time! It hasn't initially clicked with me too much the first few times. Now I think I'm liking them more.

I'm also going to try making less opinionated decisions during the migration, on my own behalf. I shouldn't be formatting things, that's not the goal here. I can be doing that either at a much later date, or just not at all. It's also not going to help with discerning changes I made in commits over time, and that's the exact issue I had in coming back to this one. I'm having a hard time figuring out when the tests stopped working, and letting that happen has gotten us to where we are now. They worked before, so I just have to ensure they keep working.