TS CJS Jest Testing, Done Right!
Heck yeah, man!!! Figured out how to get all tests passing, just by instead running them on the built version of the TS code, rather than the transpiled version of it. It's not as ideal as having it just work, but I can do that later, once things are migrated properly.

Now the built TS source is put to the `./dist` folder as well, which is where the tests/mocks run from now instead. Yoo!!!

https://github.com/jestjs/jest/issues/13143 (Kind of helped a bit, but still ran into other issues, so I went with this route instead)