Smoother Class Types
Added some missing types, as well as smoothed over some existing ones.

I almost got `XhrFileReader` working as well, but I'm not sure how to import `WorkerGlobalScope` as a type, without adding it to the global scope of the whole project itself, which doesn't seem to work to well with the rest of the global scope usage.