Missing Constructor Types
Now onto Storm Corrosion. Lock Howl really is sick! This album was from last year as well. Now I want to try making dark music like that as well, it's been a moment since I've done that.

I think this whole shebang of working on this again tonight, started with wanting to look into Art-Gen again, because I wanted to try the googleapis part again, of trying to automatically upload my videos to YouTube, after having made them with ffmpeg.

https://www.npmjs.com/package/@googleapis/youtube
https://googleapis.dev/nodejs/googleapis/latest/youtube/classes/Youtube.html
https://github.com/googleapis/google-api-nodejs-client

I think that was also spurred from this video, which was really great!
https://www.youtube.com/watch?v=jAMegKEetx4