Migration Revert
Started working tonight on moving things over to TS, but more safely, by keeping the tests passing while doing the migration, to ensure things are moved over to a more modern API, but not at the loss of ensuring things work the exact same as before.

This brings things back to the commit at the start of the migration, last year.
04473dc

https://stackoverflow.com/questions/62731388/how-to-revert-a-commit-without-losing-the-changes
https://stackoverflow.com/questions/48802948/checkout-commit-keep-changes-in-working-dir
https://stackoverflow.com/questions/1463340/how-can-i-revert-multiple-git-commits

```sh
git revert --no-commit bb44694^...HEAD
```